third_party/WebKit/LayoutTests/battery-status/detached-no-crash.html

<!DOCTYPE HTML>
<body>
<head>
<script src="../resources/js-test.js"></script>
</head>
<script>
description("Accessing navigator.getBattery() on a closed window.");

window.jsTestIsAsync = true;

var w;
var nav;

function processMessage(event) {
    if (event.data == "opened") {
        nav = w.navigator;
        shouldBeTrue("typeof(nav.getBattery()) == 'object'");
        w.close();
        w = null;
    } else if (event.data == "closed") {
      nav.getBattery().then(battery => {
        assert_unreachable('getBattery should reject on a closed window');
      })
      .catch(error => {
        // DOMException.SECURITY_ERR = 18.
        debug('Error Code is ' + error.code);
        assert_equals(error.code, DOMException.SECURITY_ERR);
      });
      setTimeout(finishJSTest, 0);
    }
}

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
}
w = window.open('../resources/window-postmessage-open-close.html');
window.addEventListener("message", processMessage, false);
</script>
</body>
